Add tests, dagcorator class
Added a decorator class (DAGCorator) that in principle can be used to add
functions to the DAG layer using a Python decorator syntax.  This has not been
extensively tested, so unsure how well it works as of now.

However added a bunch of unit tests, checking those in as well.